+++ includes/class-veritate-fact-check-crawler-activator.php
@@ -2,7 +2,7 @@
 /**
  * Fired during plugin activation
  *
- * @link       https://veritate.wowperations.com.br
+ * @link       https://veritatecrawler.wowperations.com.br
  * @since      0.1.0
  *
  * @package    Veritate_Fact_Check_Crawler