+++ includes/class-veritate-fact-check-crawler-utilities.php
@@ -5,7 +5,7 @@
  * A class definition that includes utility attributes and functions used across both the
  * public-facing side of the site and the admin area.
  *
- * @link       https://veritate.wowperations.com.br
+ * @link       https://veritatecrawler.wowperations.com.br
  * @since      0.1.0
  *
  * @package    Veritate_Fact_Check_Crawler