+++ includes/class-veritate-fact-check-crawler.php
@@ -1,12 +1,11 @@
 <?php
-
 /**
  * The file that defines the core plugin class
  *
  * A class definition that includes attributes and functions used across both the
  * public-facing side of the site and the admin area.
  *
- * @link       https://veritate.wowperations.com.br
+ * @link       https://veritatecrawler.wowperations.com.br
  * @since      0.1.0
  *
  * @package    Veritate_Fact_Check_Crawler
@@ -58,6 +57,24 @@ class Veritate_Fact_Check_Crawler {
 	protected $version;
 
 	/**
+	 * An instance of the plugin's utilities
+	 *
+	 * @since    0.1.0
+	 * @access   protected
+	 * @var      string    $utilities    An instance of the plugin's utilities
+	 */
+	protected $utilities;
+
+	/**
+	 * An instance of plugin's common class
+	 *
+	 * @since    0.1.0
+	 * @access   protected
+	 * @var      string    $common   An instance of plugin's common class
+	 */
+	protected $common;
+
+	/**
 	 * Define the core functionality of the plugin.
 	 *
 	 * Set the plugin name and the plugin version that can be used throughout the plugin.
@@ -69,7 +86,7 @@ class Veritate_Fact_Check_Crawler {
 	public function __construct() {
 
 		$this->plugin_name = 'veritate-fact-check-crawler';
-		$this->version = '0.1.0';
+		$this->version = '0.3.0';
 
 		$this->load_dependencies();
 		$this->set_locale();
@@ -85,6 +102,7 @@ class Veritate_Fact_Check_Crawler {
 	 *
 	 * - Veritate_Fact_Check_Crawler_Loader. Orchestrates the hooks of the plugin.
 	 * - Veritate_Fact_Check_Crawler_i18n. Defines internationalization functionality.
+	 * - Veritate_Fact_Check_Crawler_Utilities. Defines the all plugin utilities used on back and front ends.
 	 * - Veritate_Fact_Check_Crawler_Admin. Defines all hooks for the admin area.
 	 * - Veritate_Fact_Check_Crawler_Public. Defines all hooks for the public side of the site.
 	 *
@@ -109,9 +127,14 @@ class Veritate_Fact_Check_Crawler {
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-veritate-fact-check-crawler-i18n.php';
 
 		/**
-		 * The class responsible for defining all utilites used in both public and admin areas.
+		 * The class responsible for defining all utilities used in both public and admin areas.
 		 */
-		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'admin/class-veritate-fact-check-crawler-admin.php';
+		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-veritate-fact-check-crawler-utilities.php';
+
+		/**
+		 * The class responsible for common functionalites used in both public and admin areas.
+		 */
+		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/class-veritate-fact-check-crawler-common.php';
 
 		/**
 		 * The class responsible for defining all actions that occur in the admin area.
@@ -123,6 +146,10 @@ class Veritate_Fact_Check_Crawler {
 		 * side of the site.
 		 */
 		require_once plugin_dir_path( dirname( __FILE__ ) ) . 'public/class-veritate-fact-check-crawler-public.php';
+
+		$this->utilities = new Veritate_Fact_Check_Crawler_Utilities( $this->get_plugin_name(), $this->get_version() );
+
+		$this->common = new Veritate_Fact_Check_Crawler_Common( $this->get_plugin_name(), $this->get_version() );
 
 		$this->loader = new Veritate_Fact_Check_Crawler_Loader();
 
@@ -154,11 +181,9 @@ class Veritate_Fact_Check_Crawler {
 	 */
 	private function define_admin_hooks() {
 
-		$plugin_admin = new Veritate_Fact_Check_Crawler_Admin( $this->get_plugin_name(), $this->get_version() );
+		$plugin_admin = new Veritate_Fact_Check_Crawler_Admin( $this->get_plugin_name(), $this->get_version(), $this->get_utilities(), $this->get_common() );
 
-		$this->loader->add_action( 'admin_enqueue_scripts', $plugin_admin, 'enqueue_styles' );
-		$this->loader->add_action( 'admin_enqueue_scripts', $plugin_admin, 'enqueue_scripts' );
-
+		$this->loader->add_action( 'rest_api_init', $plugin_admin, 'register_post_api_hooks' );
 	}
 
 	/**
@@ -170,11 +195,11 @@ class Veritate_Fact_Check_Crawler {
 	 */
 	private function define_public_hooks() {
 
-		$plugin_public = new Veritate_Fact_Check_Crawler_Public( $this->get_plugin_name(), $this->get_version() );
+		$plugin_public = new Veritate_Fact_Check_Crawler_Public( $this->get_plugin_name(), $this->get_version(), $this->get_utilities(), $this->get_common() );
 
-		$this->loader->add_action( 'wp_enqueue_scripts', $plugin_public, 'enqueue_styles' );
-		$this->loader->add_action( 'wp_enqueue_scripts', $plugin_public, 'enqueue_scripts' );
-
+		$this->loader->add_action( 'template_redirect', $plugin_public, 'veritate_headless_redirect', 20 );
+		$this->loader->add_action( 'rest_endpoints', $plugin_public, 'veritate_rest_endpoints' );
+		$this->loader->add_filter( 'rest_prepare_post', $plugin_public, 'rest_prepare_post', 10, 3 );
 	}
 
 	/**
@@ -215,6 +240,26 @@ class Veritate_Fact_Check_Crawler {
 	 */
 	public function get_version() {
 		return $this->version;
+	}
+
+	/**
+	 * Retrieve the plugin utitlies.
+	 *
+	 * @since     0.1.0
+	 * @return    object    Plugin utilities
+	 */
+	public function get_utilities() {
+		return $this->utilities;
+	}
+
+	/**
+	 * Retrieve the plugin common functions and methods
+	 *
+	 * @since     0.1.0
+	 * @return    object   An instance of common class holding methods and properties used in front and backend
+	 */
+	public function get_common() {
+		return $this->common;
 	}
 
 }