+++ public/class-veritate-fact-check-crawler-public.php
@@ -1,9 +1,8 @@
 <?php
-
 /**
  * The public-facing functionality of the plugin.
  *
- * @link       https://veritate.wowperations.com.br
+ * @link       https://veritatecrawler.wowperations.com.br
  * @since      0.1.0
  *
  * @package    Veritate_Fact_Check_Crawler
@@ -41,63 +40,125 @@ class Veritate_Fact_Check_Crawler_Public {
 	private $version;
 
 	/**
+	 * An instance of the plugin's utilities
+	 *
+	 * @since    0.3.0
+	 * @access   protected
+	 * @var      string    $utilities    An instance of the plugin's utilities.
+	 */
+	protected $utilities;
+
+	/**
+	 * An instance of plugin's common class
+	 *
+	 * @since    0.3.0
+	 * @access   protected
+	 * @var      string    $common   An instance of plugin's common class.
+	 */
+	protected $common;
+
+	/**
 	 * Initialize the class and set its properties.
 	 *
-	 * @since    0.1.0
-	 * @param      string    $plugin_name       The name of the plugin.
-	 * @param      string    $version    The version of this plugin.
+	 * @since 0.1.0
+	 * @param string $plugin_name       The name of this plugin.
+	 * @param string $version    The version of this plugin.
+	 * @since 0.3.0
+	 * @param object $utilities An instance of the plugin's utilities.
+	 * @param object $common    An instance of plugin's common class.
 	 */
-	public function __construct( $plugin_name, $version ) {
+	public function __construct( $plugin_name, $version, $utilities, $common ) {
 
 		$this->plugin_name = $plugin_name;
 		$this->version = $version;
+		$this->utilities = $utilities;
+		$this->common = $common;
 
 	}
 
 	/**
-	 * Register the stylesheets for the public-facing side of the site.
+	 * Redirect All WordPress frontend requests to the home
 	 *
-	 * @since    0.1.0
+	 * Redirect All WordPress frontend requests to home, except for WP Rest API endpoints.
+	 * This function is also declared in Veritate API Theme as fallback and these redirection
+	 * will also be declared in .htaccess/.htconfig and in reverse-proxy layer (Nginx or Cloudflare)
+	 *
+	 * @since 0.1.0
+	 * @since 0.2.0 uses wp_safe_redirect, uses 301 redirection and don't redirect for 404 requests.
 	 */
-	public function enqueue_styles() {
-
-		/**
-		 * This function is provided for demonstration purposes only.
-		 *
-		 * An instance of this class should be passed to the run() function
-		 * defined in Veritate_Fact_Check_Crawler_Loader as all of the hooks are defined
-		 * in that particular class.
-		 *
-		 * The Veritate_Fact_Check_Crawler_Loader will then create the relationship
-		 * between the defined hooks and the functions defined in this
-		 * class.
-		 */
-
-		wp_enqueue_style( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'css/veritate-fact-check-crawler-public.css', array(), $this->version, 'all' );
-
+	public function veritate_headless_redirect() {
+		if ( ! is_front_page() ) {
+			wp_safe_redirect( home_url(), 301 );
+			exit;
+		}
 	}
 
 	/**
-	 * Register the JavaScript for the public-facing side of the site.
+	 * Disable undesired WP Rest API Endpoints
 	 *
-	 * @since    0.1.0
+	 * Disables some unused/undesired WP Rest API endpoints.
+	 * This function is also declared in Veritate API Theme as fallback
+	 *
+	 * @since  0.1.0
+	 * @param  array $endpoints an array with all registered REST API Endpoints.
+	 *
+	 * @return array $endpoints an filtered array of all registered REST API Endpoints.
 	 */
-	public function enqueue_scripts() {
+	public function veritate_rest_endpoints( $endpoints ) {
+		if ( isset( $endpoints['/wp/v2/users'] ) ) {
+			unset( $endpoints['/wp/v2/users'] );
+		}
+		if ( isset( $endpoints['/wp/v2/users/(?P<id>[\d]+)'] ) ) {
+			unset( $endpoints['/wp/v2/users/(?P<id>[\d]+)'] );
+		}
+		if ( isset( $endpoints['/wp/v2/pages'] ) ) {
+			unset( $endpoints['/wp/v2/pages'] );
+		}
+		if ( isset( $endpoints['/wp/v2/pages/(?P<id>[\d]+)'] ) ) {
+			unset( $endpoints['/wp/v2/pages/(?P<id>[\d]+)'] );
+		}
+		if ( isset( $endpoints['/wp/v2/media'] ) ) {
+			unset( $endpoints['/wp/v2/media'] );
+		}
+		if ( isset( $endpoints['/wp/v2/media/(?P<id>[\d]+)'] ) ) {
+			unset( $endpoints['/wp/v2/media/(?P<id>[\d]+)'] );
+		}
+		if ( isset( $endpoints['/wp/v2/types'] ) ) {
+			unset( $endpoints['/wp/v2/types'] );
+		}
+		if ( isset( $endpoints['/wp/v2/types/(?P<id>[\d]+)'] ) ) {
+			unset( $endpoints['/wp/v2/types/(?P<id>[\d]+)'] );
+		}
+		if ( isset( $endpoints['/wp/v2/statuses'] ) ) {
+			unset( $endpoints['/wp/v2/statuses'] );
+		}
+		if ( isset( $endpoints['/wp/v2/statuses/(?P<id>[\d]+)'] ) ) {
+			unset( $endpoints['/wp/v2/statuses/(?P<id>[\d]+)'] );
+		}
+		if ( isset( $endpoints['/wp/v2/comments'] ) ) {
+			unset( $endpoints['/wp/v2/comments'] );
+		}
+		if ( isset( $endpoints['/wp/v2/comments/(?P<id>[\d]+)'] ) ) {
+			unset( $endpoints['/wp/v2/comments/(?P<id>[\d]+)'] );
+		}
+		return $endpoints;
+	}
 
-		/**
-		 * This function is provided for demonstration purposes only.
-		 *
-		 * An instance of this class should be passed to the run() function
-		 * defined in Veritate_Fact_Check_Crawler_Loader as all of the hooks are defined
-		 * in that particular class.
-		 *
-		 * The Veritate_Fact_Check_Crawler_Loader will then create the relationship
-		 * between the defined hooks and the functions defined in this
-		 * class.
-		 */
-
-		wp_enqueue_script( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'js/veritate-fact-check-crawler-public.js', array( 'jquery' ), $this->version, false );
-
+	/**
+	 * Filter post REST response using Veritate_Fact_Check_Crawler_Common::rest_prepare_post.
+	 *
+	 * A wrapper for Veritate_Fact_Check_Crawler_Common::rest_prepare_post, which filters
+	 * post REST API response .
+	 *
+	 * @since 0.3.0
+	 * @access public
+	 * @param object $data    The response object.
+	 * @param object $post    The original post type object.
+	 * @param object $request Request used to generate the response.
+	 * @return object $data The filtered response object.
+	 */
+	public function rest_prepare_post( $data, $post, $request ) {
+		return $this->common->rest_prepare_post( $data, $post, $request );
 	}
 
 }